refactor component file name
